+++ eslint-bridge/profiling/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "server": "node --inspect-brk server.js",
     "compile-server": "cd ..; npm run compile; cd profiling; node --inspect-brk server.js",
-    "analyze-js": "../node_modules/.bin/ts-node analyze-js.ts"
+    "profile": "../node_modules/.bin/ts-node profile-rule.ts"
   },
   "license": "LGPL-3.0",
   "devDependencies": {
