+++ eslint-bridge/profiling/server.js
@@ -22,11 +22,12 @@
 const server = require('../lib/server');
 const path = require('path');
 const context = require('../lib/helpers');
+const { tmpdir } = require('os');
 
-// must be the same as the one used in ./analyze.ts
+// must be the same as the one used in ./profile-rule.ts
 const port = 64829;
 const host = '127.0.0.1';
-const workDir = '/tmp/dir';
+const workDir = tmpdir();
 
 context.setContext({ workDir, shouldUseTypeScriptParserForJS: false, sonarlint: false, bundles: [] });
 const BIG_TIMEOUT = 1719925474;
